update mvc, conneciton pool, redis
