+++ models/attraction.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class Attraction(BaseModel):
+	id: int
+	name: str
+	address: str
+	image: str
