+++ models/token.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
