+++ static/booking.html
@@ -74,9 +74,9 @@
   </div>
   <div class="booking-container">
     <div class="contact-info">您的聯絡資訊</div>
-    <div class="input-group">聯絡姓名：<input type="naem" placeholder="name" class="input-name"></div>
-    <div class="input-group">連絡信箱：<input type="email" placeholder="輸入電子郵件" class="input-mail"></div>
-    <div class="input-group">手機號碼：<input type="number" placeholder="輸入手機號碼" class="input-number"></div>
+      <div class="input-group">聯絡姓名：<input type="name" placeholder="name" class="input-name"></div>
+      <div class="input-group">連絡信箱：<input type="email" placeholder="輸入電子郵件" class="input-mail"></div>
+      <div class="input-group">手機號碼：<input type="number" placeholder="輸入手機號碼" class="input-number"></div>
     <div class="notice">請保持手機暢通，準時到達，導覽人員將用手機與您聯繫，務必留下正確的聯絡方式。</div>
   </div>  
   <div class="horizontal-line-container">
@@ -84,7 +84,7 @@
   </div>
   <div class="booking-container">
     <div class="creditcard-info">信用卡付款資訊</div>
-    <form id="payment-form">
+    <form id="payment-form-creditcard">
     <div class="input-group">
       卡片號碼：<div class="credit-card-input tpfield" id="card-number"></div>
     </div>
